Updated files module
