+++ pans_labyrinth/__init__.py
@@ -1,0 +1,1 @@
+__version__ = "0.0.1"
